==> 2_data_preprocessinge.ipynb <==
# 1 Merged gene expression matrix and disease-free survival (DFS).

import numpy as np
import pandas as pd

# Load EDASeq normalized data.
allPatients = pd.read_csv('data/EDASeq_Normalized.csv')
allPatients

allPatients = allPatients.T
allPatients.columns = allPatients.iloc[0, :]
allPatients.insert(0, 'submitter', allPatients.index)
allPatients

# Process patient ID

allPatients.iloc[1:, 0] = allPatients.iloc[1:, 0].map(lambda x : '-'.join(x.split('-')[:4]))
allPatients

# Read DFS data.

all_dfs = pd.read_excel("data/DFS.xlsx", usecols=[0, 4])
all_dfs = all_dfs.rename(columns={'submitter_id.samples': 'submitter'})
all_dfs

# Merge gene expression matrix and DFS according to patient ID.

myDataset = pd.merge(allPatients, all_dfs, how='inner', on='submitter')
myDataset

# Check the merged genes.

gene = myDataset.columns[1 : -1].to_numpy()
print(gene.shape)
print(gene)

# Check the merged gene expression matrix and DFS.

data = myDataset.drop('submitter', axis = 1).to_numpy()

print(f'shape = {data.shape}')
print(data)

# 2 Use 12 months to divide radiation resistance (RR) and radiation sensitivity (RS).
+ Samples from 9-15 months have been manually removed.
+ RR: DFS is less than or equal to 12 months.
+ RS: DFS greater than 12 months.

data[:, -1:] = np.where(data[:, -1:] > 12, 0, 1)
print(data.shape)
print(data)

feature = data[:, :-1].astype('float')
label = data[:, -1:].astype('int64')
print(feature.shape)
print(label.shape)

# 3 Genes with an expression level of 0 were removed.

# For each gene, if there is a patient with an expression level of 0 for that gene, the gene is removed.

reservedList = []

for i in range(feature.shape[1]):
    if feature[:, i].tolist().count(0) == 0: # reserve
        reservedList.append(i)

feature = feature[:, reservedList]
gene = gene[reservedList]
print(feature.shape)
print(gene.shape)
print(gene)
print(feature)
np.save('data/geneAfterDiscard_0.npy', gene)

# 4 Oversampling of radiotherapy-resistant samples using the SMOTE algorithm.

from imblearn.over_sampling import SMOTE
from collections import Counter

print(f'before overstamp :{Counter(label.reshape(-1))}')

# Set random state for recurring.

overstamp = SMOTE(random_state=2022)
x, y = overstamp.fit_resample(feature, label)

print(f'after overstamp :{Counter(y)}')

y = y.reshape((-1, 1))
oversampled_data = np.append(x, y, axis = 1)

np.save("data/oversampled_data.npy", oversampled_data)
