==> 5.2_classifier_comparison_9_genes.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt

data = np.load('data/oversampled_data.npy', allow_pickle=True)
gene = np.load('data/geneAfterDiscard_0.npy', allow_pickle=True)

feature = data[:, :-1]
label = data[:, -1]
print(gene.shape)
print(feature.shape)
print(label.shape)

# Take out the gene of choice.

fina_gene = ['SPAG1', 'FIGN', 'NUBPL', 'CHMP5', 'TCF7L2', 'COQ10B', 'BSDC1', 'ZFPM1', 'GRPEL1']
gene_idx = []
for g in fina_gene:
    gene_idx.append(np.where(gene==g)[0][0])
gene_idx = np.array(gene_idx)

fina_feature = feature[:, gene_idx]

from sklearn.model_selection import train_test_split
from sklearn.metrics import accuracy_score
from sklearn.metrics import precision_score
from sklearn.metrics import recall_score
from sklearn.metrics import roc_auc_score

import time

X_train, X_test, y_train, y_test = train_test_split(fina_feature, label, test_size=0.3, shuffle=True, random_state=2022)

# RandomForestClassifier

start = time.time()
from sklearn.ensemble import RandomForestClassifier

allAcc = []
allPrec = []
allRec = []
allAuc = []
for _ in range(5):
    clf = RandomForestClassifier()
    clf.fit(X_train, y_train)  
    y_pred = clf.predict(X_test)

    allAcc.append(accuracy_score(y_test, y_pred))
    allPrec.append(precision_score(y_test, y_pred))
    allRec.append(recall_score(y_test, y_pred))
    predict_prob_y = clf.predict_proba(X_test)[:, 1]
    allAuc.append(roc_auc_score(y_test, predict_prob_y))

end = time.time()

t = end -start
print(f'{round(np.mean(allAcc), 4)}\t{round(np.mean(allPrec), 4)}\t{round(np.mean(allRec), 4)}\t{round(np.mean(allAuc), 4)}\t{t}')

# GradientBoostingClassifier

start = time.time()
from sklearn.ensemble import GradientBoostingClassifier

allAcc = []
allPrec = []
allRec = []
allAuc = []
for _ in range(5):
    clf = GradientBoostingClassifier()
    clf.fit(X_train, y_train)  
    y_pred = clf.predict(X_test)

    allAcc.append(accuracy_score(y_test, y_pred))
    allPrec.append(precision_score(y_test, y_pred))
    allRec.append(recall_score(y_test, y_pred))
    predict_prob_y = clf.predict_proba(X_test)[:, 1]
    allAuc.append(roc_auc_score(y_test, predict_prob_y))

end = time.time()

t = end -start
print(f'{round(np.mean(allAcc), 4)}\t{round(np.mean(allPrec), 4)}\t{round(np.mean(allRec), 4)}\t{round(np.mean(allAuc), 4)}\t{t}')

# svm

start = time.time()
from sklearn import svm

allAcc = []
allPrec = []
allRec = []
allAuc = []
for _ in range(5):
    clf = svm.SVC(probability=True)
    clf.fit(X_train, y_train)  
    y_pred = clf.predict(X_test)

    allAcc.append(accuracy_score(y_test, y_pred))
    allPrec.append(precision_score(y_test, y_pred))
    allRec.append(recall_score(y_test, y_pred))
    predict_prob_y = clf.predict_proba(X_test)[:, 1]
    allAuc.append(roc_auc_score(y_test, predict_prob_y))

end = time.time()

t = end -start
print(f'{round(np.mean(allAcc), 4)}\t{round(np.mean(allPrec), 4)}\t{round(np.mean(allRec), 4)}\t{round(np.mean(allAuc), 4)}\t{t}')

# LogisticRegression

start = time.time()
from sklearn.linear_model import LogisticRegression

allAcc = []
allPrec = []
allRec = []
allAuc = []
for _ in range(5):
    clf = LogisticRegression()
    clf.fit(X_train, y_train)  
    y_pred = clf.predict(X_test)

    allAcc.append(accuracy_score(y_test, y_pred))
    allPrec.append(precision_score(y_test, y_pred))
    allRec.append(recall_score(y_test, y_pred))
    predict_prob_y = clf.predict_proba(X_test)[:, 1]
    allAuc.append(roc_auc_score(y_test, predict_prob_y))

end = time.time()

t = end -start
print(f'{round(np.mean(allAcc), 4)}\t{round(np.mean(allPrec), 4)}\t{round(np.mean(allRec), 4)}\t{round(np.mean(allAuc), 4)}\t{t}')

# KNeighborsClassifier

start = time.time()
from sklearn.neighbors import KNeighborsClassifier

allAcc = []
allPrec = []
allRec = []
allAuc = []
for _ in range(5):
    clf = KNeighborsClassifier()
    clf.fit(X_train, y_train)  
    y_pred = clf.predict(X_test)

    allAcc.append(accuracy_score(y_test, y_pred))
    allPrec.append(precision_score(y_test, y_pred))
    allRec.append(recall_score(y_test, y_pred))
    predict_prob_y = clf.predict_proba(X_test)[:, 1]
    allAuc.append(roc_auc_score(y_test, predict_prob_y))

end = time.time()

t = end -start
print(f'{round(np.mean(allAcc), 4)}\t{round(np.mean(allPrec), 4)}\t{round(np.mean(allRec), 4)}\t{round(np.mean(allAuc), 4)}\t{t}')

# MLPClassifier

start = time.time()
from sklearn.neural_network import MLPClassifier

allAcc = []
allPrec = []
allRec = []
allAuc = []
for _ in range(5):
    clf = MLPClassifier(hidden_layer_sizes=(180, 180, 60, 50), max_iter=300)
    clf.fit(X_train, y_train)  
    y_pred = clf.predict(X_test)

    allAcc.append(accuracy_score(y_test, y_pred))
    allPrec.append(precision_score(y_test, y_pred))
    allRec.append(recall_score(y_test, y_pred))
    predict_prob_y = clf.predict_proba(X_test)[:, 1]
    allAuc.append(roc_auc_score(y_test, predict_prob_y))

end = time.time()

t = end -start
print(f'{round(np.mean(allAcc), 4)}\t{round(np.mean(allPrec), 4)}\t{round(np.mean(allRec), 4)}\t{round(np.mean(allAuc), 4)}\t{t}')
